backend/requirements.txt
anyio==3.7.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
click==8.1.3
dnspython==2.3.0
docx==0.2.4
email-validator==2.0.0.post2
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fastapi==0.98.0
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.3
numpy==1.25.0
openpyxl==3.1.2
orjson==3.9.1
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pydantic==1.10.9
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
tesseract==0.1.3
typing_extensions==4.6.3
tzdata==2023.3
ujson==5.8.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
XlsxWriter==3.1.2
